requirements.txt
a2wsgi==1.10.8
absl-py==2.0.0
aiologic==0.14.0
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.6.0
ansicolors==1.1.8
anyio==4.1.0
apache-airflow==3.0.0
apache-airflow-core==3.0.0
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-papermill==3.10.1
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-sqlite==3.6.0
apache-airflow-providers-standard==1.1.0
apache-airflow-task-sdk==1.0.0
apispec==6.3.0
argcomplete==3.2.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
av==14.3.0
azure-core==1.34.0
azure-datalake-store==0.0.53
azure-identity==1.21.0
azure-storage-blob==12.25.1
Babel==2.14.0
backoff==2.2.1
black==25.1.0
blinker==1.7.0
bottle==0.13.3
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.3.2
cadwyn==5.3.3
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
clickclick==20.10.2
clr_loader==0.2.7.post0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
ConfigUpdater==3.2
confluent-kafka==2.6.0
connexion==2.14.2
cron-descriptor==1.4.0
croniter==6.0.0
cryptography==41.0.7
cssselect2==0.8.0
ctranslate2==4.6.0
dacite==1.9.2
deepface==0.0.93
Deprecated==1.2.14
dill==0.3.1.1
dm-tree==0.1.8
dnspython==2.4.2
docutils==0.20.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
faster-whisper==1.1.1
filelock==3.18.0
fire==0.7.0
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2025.3.2
gcsfs==2025.3.2
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-pasta==0.2.0
google-re2==1.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.1
greenlet==3.0.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==21.2.0
h11==0.16.0
htmlmin==0.1.12
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.6
ImageHash==4.3.1
importlib-metadata==6.11.0
importlib-resources==6.1.1
imutils==0.5.4
ipykernel==6.29.5
isodate==0.7.2
itsdangerous==2.1.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kaleido==0.2.1
lazy-object-proxy==1.10.0
libclang==16.0.6
libcst==1.7.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
lz4==4.4.3
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.3
ml-dtypes==0.2.0
mne==1.0.0
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msgspec==0.19.0
mtcnn==1.0.0
multidict==6.0.4
multimethod==1.12
namex==0.0.7
nbconvert==7.16.6
networkx==3.4.2
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.22.0
opencv-contrib-python==4.9.0.80
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.18
packaging==23.2
pandas==2.1.4
papermill==2.6.0
pathspec==0.12.1
pendulum==2.1.2
pgmpy==0.1.24
phik==0.12.4
pillow==11.2.1
pluggy==1.5.0
pooch==1.8.1
prison==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
proxy-tools==0.1.0
psutil==7.0.0
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydyf==0.11.0
PyGithub==2.6.1
Pygments==2.17.2
PyJWT==2.10.1
pyparsing==3.2.3
pyphen==0.17.2
pyreadline3==3.5.4
python-daemon==3.0.1
python-dotenv==1.1.0
python-multipart==0.0.20
python-nvd3==0.15.0
python-slugify==8.0.1
pythonnet==3.0.5
pytz==2023.3.post1
pytzdata==2020.1
pywebview==5.4
PyYAML==6.0.1
pyzmq==26.4.0
rarfile==4.2
referencing==0.32.0
requests==2.32.3
requests-oauthlib==1.3.1
retina-face==0.0.17
retryhttp==1.3.3
rfc3986==1.5.0
rich==14.0.0
rich-argparse==1.4.0
rich-toolkit==0.14.4
rpds-py==0.13.2
rsa==4.9
scrapbook==0.5.0
setproctitle==1.3.3
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
ssqueezepy==0.6.5
starlette==0.46.2
structlog==25.3.0
svcs==25.1.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tkwebview2==3.5.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
typeguard==4.4.2
typer==0.15.3
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.2
ujson==5.10.0
unicodecsv==0.14.1
universal_pathlib==0.2.6
uritemplate==4.1.1
urllib3==1.26.20
uuid6==2024.7.10
uvicorn==0.34.2
visions==0.8.1
watchfiles==1.0.5
weasyprint==65.1
websockets==15.0.1
wirerope==1.0.0
wordcloud==1.9.4
WTForms==3.1.1
ydata-profiling==4.16.1
zipp==3.17.0
zopfli==0.2.3.post1
